refactor transfer utility functions
