+++ codesign/lib/src/utils.dart
@@ -10,85 +10,6 @@ import 'package:process/process.dart';
 
 import 'log.dart';
 
-/// Utility function to upload a file to google cloud.
-Future<void> uploadFunction({
-  required String localPath,
-  required String destinationUrl,
-  required ProcessManager processManager,
-}) async {
-  final ProcessResult result = await processManager.run(
-    <String>['gsutil', 'cp', localPath, destinationUrl],
-  );
-  if (result.exitCode != 0) {
-    throw Exception('Failed to upload $localPath to $destinationUrl');
-  }
-}
-
-/// Utility function to download a file from google cloud.
-Future<File> downloadFunction({
-  required String sourceUrl,
-  required String localPath,
-  required ProcessManager processManager,
-  required Directory rootDirectory,
-}) async {
-  final ProcessResult result = await processManager.run(
-    <String>['gsutil', 'cp', sourceUrl, localPath],
-  );
-  if (result.exitCode != 0) {
-    throw CodesignException('Failed to download from $sourceUrl');
-  }
-  return rootDirectory.fileSystem.file(localPath);
-}
-
-/// Utility function class to handle upload/download of files from/to google cloud.
-class EngineArtifactTransfer {
-  EngineArtifactTransfer({
-    required this.gsCloudBaseUrl,
-    required this.uploadFunction,
-    required this.downloadFunction,
-  });
-
-  String gsCloudBaseUrl;
-  Function uploadFunction;
-  Function downloadFunction;
-
-  /// Wrapper function to upload code signed flutter engine artifact to google cloud bucket.
-  Future<void> uploadEngineArtifact({
-    required String localPath,
-    required String remotePath,
-    required String commitHash,
-    required ProcessManager processManager,
-    int exitCode = 0,
-  }) async {
-    final String fullRemotePath = '$gsCloudBaseUrl/flutter/$commitHash/$remotePath';
-    return await uploadFunction(
-      localPath: localPath,
-      destinationUrl: fullRemotePath,
-      processManager: processManager,
-      exitCode: exitCode,
-    );
-  }
-
-  /// Wrapper function to download flutter engine artifact from google cloud bucket.
-  Future<File> downloadEngineArtifact({
-    required String remotePath,
-    required String localPath,
-    required String commitHash,
-    required ProcessManager processManager,
-    required Directory rootDirectory,
-    int exitCode = 0,
-  }) async {
-    final String sourceUrl = '$gsCloudBaseUrl/flutter/$commitHash/$remotePath';
-    return await downloadFunction(
-      sourceUrl: sourceUrl,
-      localPath: localPath,
-      processManager: processManager,
-      rootDirectory: rootDirectory,
-      exitCode: exitCode,
-    );
-  }
-}
-
 enum FileType {
   folder,
   zip,
